perf(course-page-edit): coalesce insert panel preview updates
